Validações de remoção de teste e validações do cadastro de teste

+++ src/iris/service/TesteService.java
@@ -17,8 +17,10 @@ import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
+import javax.ws.rs.WebApplicationException;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
+import javax.ws.rs.core.Response.Status;
 
 import org.json.JSONException;
 
@@ -119,6 +121,9 @@ public class TesteService {
 	@Path("/{testeId}")
 	@Produces(MediaType.APPLICATION_JSON)
 	public void removerTeste(@PathParam("testeId") final int testeId) {
+		if(testeDB.hasResults(testeId)){
+			throw new WebApplicationException(Status.BAD_REQUEST);
+		}
 		testeDB.delete(testeId);
 	}
